requirements.txt
asgiref==3.7.2
Django==4.2.9
django-admin-interface==0.28.1
django-colorfield==0.11.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
pillow==10.2.0
psycopg==3.1.16
python-slugify==8.0.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
